+++ src/constraint/url-protocol.ts
@@ -5,13 +5,16 @@ export function createURLProtocolConstraint(
   protocols: string[],
 ): ExtrinsicConstraint<URL> {
   const listFormatter = createDisjunctionFormatter();
-  const description = `protocol must be ${listFormatter.format(protocols)}`;
+  const inlineCodeProtocols = protocols.map((protocol) => `\`${protocol}\``);
 
   return {
-    description,
+    description: `protocol must be ${listFormatter.format(inlineCodeProtocols)}`,
     constrain: function constrainURLProtocol({ protocol }) {
       if (!protocols.includes(protocol)) {
-        return `${description}, but is ${protocol}`;
+        return (
+          `protocol must be ${listFormatter.format(protocols)}, ` +
+          `but is ${protocol}`
+        );
       }
     },
   };